New branch with separation for task_hmm
